+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Release 0.0.16 (Unreleased)
+## Release 0.1.0 (Unreleased)
 
 ### Added